Extract base class IdentityUserBase from IdentityUser to allow different subclasses that support non-ObjectId Ids.
