--- src/CoreTests/project.json
@@ -1,26 +1,0 @@
-{
-  "version": "1.0.0-*",
-
-  "testRunner": "nunit",
-
-  "dependencies": {
-    "dotnet-test-nunit": "3.4.0-beta-2",
-    "Microsoft.AspNetCore.Identity.MongoDB": "1.0.0-*",
-    "Microsoft.Extensions.DependencyInjection": "1.0.0",
-    "Microsoft.Extensions.Logging": "1.0.0",
-    "NUnit": "3.4.1"
-  },
-
-  "frameworks": {
-    "net451": {},
-    "netcoreapp1.0": {
-      "imports": "portable-net45+win8",
-      "dependencies": {
-        "Microsoft.NETCore.App": {
-          "version": "1.0.0-*",
-          "type": "platform"
-        }
-      }
-    }
-  }
-}
