+++ src/Microsoft.AspNetCore.Identity.MongoDB/IndexChecks.cs
@@ -5,7 +5,7 @@
 	public static class IndexChecks
 	{
 		public static void EnsureUniqueIndexOnNormalizedUserName<TUser>(IMongoCollection<TUser> users)
-			where TUser : IdentityUser
+			where TUser : IdentityUserBase
 		{
 			var userName = Builders<TUser>.IndexKeys.Ascending(t => t.NormalizedUserName);
 			var unique = new CreateIndexOptions {Unique = true};
@@ -21,7 +21,7 @@
 		}
 
 		public static void EnsureUniqueIndexOnNormalizedEmail<TUser>(IMongoCollection<TUser> users)
-			where TUser : IdentityUser
+			where TUser : IdentityUserBase
 		{
 			var email = Builders<TUser>.IndexKeys.Ascending(t => t.NormalizedEmail);
 			var unique = new CreateIndexOptions {Unique = true};
@@ -35,8 +35,8 @@
 		public static class OptionalIndexChecks
 		{
 			public static void EnsureUniqueIndexOnUserName<TUser>(IMongoCollection<TUser> users)
-				where TUser : IdentityUser
-			{
+				where TUser : IdentityUserBase
+            {
 				var userName = Builders<TUser>.IndexKeys.Ascending(t => t.UserName);
 				var unique = new CreateIndexOptions {Unique = true};
 				users.Indexes.CreateOneAsync(userName, unique);
@@ -44,15 +44,15 @@
 
 			public static void EnsureUniqueIndexOnRoleName<TRole>(IMongoCollection<TRole> roles)
 				where TRole : IdentityRole
-			{
+            {
 				var roleName = Builders<TRole>.IndexKeys.Ascending(t => t.Name);
 				var unique = new CreateIndexOptions {Unique = true};
 				roles.Indexes.CreateOneAsync(roleName, unique);
 			}
 
 			public static void EnsureUniqueIndexOnEmail<TUser>(IMongoCollection<TUser> users)
-				where TUser : IdentityUser
-			{
+				where TUser : IdentityUserBase
+            {
 				var email = Builders<TUser>.IndexKeys.Ascending(t => t.Email);
 				var unique = new CreateIndexOptions {Unique = true};
 				users.Indexes.CreateOneAsync(email, unique);
