+++ src/Microsoft.AspNetCore.Identity.MongoDB/Microsoft.AspNetCore.Identity.MongoDB.csproj
@@ -1,0 +1,36 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <Description>A MongoDB provider for ASP.NET Core Identity framework.</Description>
+    <VersionPrefix>1.0.2</VersionPrefix>
+    <Authors>Wes Higbee</Authors>
+    <TargetFrameworks>net451;netstandard1.5</TargetFrameworks>
+    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
+    <AssemblyName>Microsoft.AspNetCore.Identity.MongoDB</AssemblyName>
+    <PackageId>Microsoft.AspNetCore.Identity.MongoDB</PackageId>
+    <PackageTags>aspnetcore;mongo;mongodb;identity;membership</PackageTags>
+    <PackageReleaseNotes>Extract base class IdentityUserBase from IdentityUser to allow different subclasses that support non-ObjectId Ids.</PackageReleaseNotes>
+    <PackageProjectUrl>https://github.com/g0t4/aspnet-identity-mongo</PackageProjectUrl>
+    <PackageLicenseUrl>https://github.com/g0t4/aspnet-identity-mongo/blob/master/LICENSE</PackageLicenseUrl>
+    <RepositoryType>git</RepositoryType>
+    <RepositoryUrl>https://github.com/g0t4/aspnet-identity-mongo</RepositoryUrl>
+    <NetStandardImplicitPackageVersion Condition=" '$(TargetFramework)' == 'netstandard1.5' ">1.6.0</NetStandardImplicitPackageVersion>
+    <GenerateAssemblyConfigurationAttribute>false</GenerateAssemblyConfigurationAttribute>
+    <GenerateAssemblyCompanyAttribute>false</GenerateAssemblyCompanyAttribute>
+    <GenerateAssemblyProductAttribute>false</GenerateAssemblyProductAttribute>
+    <Version>1.1.1</Version>
+    <AssemblyVersion>1.1.1.0</AssemblyVersion>
+    <FileVersion>1.1.1.0</FileVersion>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.AspNetCore.Identity" Version="1.0.2" />
+    <PackageReference Include="MongoDB.Driver" Version="2.3.0" />
+  </ItemGroup>
+
+  <ItemGroup Condition=" '$(TargetFramework)' == 'net451' ">
+    <Reference Include="System" />
+    <Reference Include="Microsoft.CSharp" />
+  </ItemGroup>
+
+</Project>
