--- src/Microsoft.AspNetCore.Identity.MongoDB/project.json
@@ -1,47 +1,0 @@
-{
-  "name": "Microsoft.AspNetCore.Identity.MongoDB",
-  "version": "1.0.2",
-
-  "description":
-    "A MongoDB provider for ASP.NET Core Identity framework.",
-
-  "authors": ["Wes Higbee"],
-
-  "buildOptions": {
-    "warningsAsErrors": true
-  },
-
-  "packOptions": {
-    "summary": "A MongoDB provider for ASP.NET Core Identity",
-    "owners": ["Wes Higbee"],
-    "releaseNotes": "Convert back to using DateTime to store LockoutEndDate, DateTimeOffset serializes to an array of values which could make it hard for people to sort on this and query on this. Also DateTime was used in the v2 driver, so this makes the upgrade story easier.",
-
-    "tags": [
-      "aspnetcore",
-      "mongo",
-      "mongodb",
-      "identity",
-      "membership"
-    ],
-    "repository": {
-      "type": "git",
-      "url": "https://github.com/g0t4/aspnet-identity-mongo"
-    },
-    "licenseUrl": "https://github.com/g0t4/aspnet-identity-mongo/blob/master/LICENSE",
-    "projectUrl": "https://github.com/g0t4/aspnet-identity-mongo"
-  },
-
-  "dependencies": {
-    "Microsoft.AspNetCore.Identity": "1.0.0",
-    "MongoDB.Driver": "2.3.0"
-  },
-
-  "frameworks": {
-    "net451": {},
-    "netstandard1.5": {
-      "dependencies": {
-        "NETStandard.Library": "1.6.0"
-      }
-    }
-  }
-}
